+++ main.py
@@ -29,6 +29,9 @@ if __name__ == "__main__":
     args = parser.parse_args()
     videos = os.listdir("dataset/test/detection/videos")
     pprint(args)
+    if args.align_face is True and args.extract_face is False:
+        print("You need to extract faces first. Alignment disabled for now.")
+
     display_video_motpy(
         filepath="dataset/test/detection/videos/" + videos[4],
         model=args.detector,
